Fix topology creation to use the residue from the hybrid topology
This problem did not affect actual system creation, but resulted in the
MDTraj topology that was generated being unable to be deepcopied. To
resolve it, for each unique new atom that gets added, we find the
corresponding residue and extract it from the hybrid, then use that to
add the new atom.