Switch to container class results for the tasklets
In order to facilitate chaining tasklets together in a fully
asynchronous manner, I have created namedtuples for equilibrium and
nonequilibrium results. This allows the appropriate tasks to receive the
object that they were expecting.

Cons: this results in sending more data than one would like over the
network. It isn't directly clear how to resolve this; if celery could
pass a subset of return values, that would resolve this issue.