Add a basic setup script that takes a yaml file as input
Previously, only the API was available. This was cumbersome and did not
make clear which options should be modified on ordinary runs. This
commit introduces a script that takes in a simple yaml file to specify
how a calculation should be set up and where its output should go. The
script also runs the calculation.