Add plotting tools library and changes to setup script
This commit adds a library for accessing the data that was written to
disk during a calculation. This facilitates viewing the data in a
jupyter notebook or elsewhere, if the user prefers.

This commit also includes a change to the setup script. We now attempt
to write out the run object as well as the setup object.