Expand dictionary of integrator options to individual keyword arguments
Now that the integrator is no longer serialized, it has to be instantiated
by the tasklet. As such, we need to pass arguments. Initially we were passing
a dictionary, but this is difficult from a user's point of view: what needs
to be in the dictionary? There were few elements in the dictionary, so that
has been converted to individual keyword arguments