Begin implementing AtomMapper
This commit begins the implementation of a utility class to map groups
of molecules.

Purpose: To modularize atom mapping, allow pre-enumeration of atom maps
(for performance), and to add extra code to verify that the maps that are
used in an RJ calculation are actually valid for geometry proposals.

This is an initial commit, so what remains includes:

1. Code that verifies an atom map
2. Code that verifies the graph of proposals (ensuring it is connected)
3. to/from JSON methods to allow pre-generation of atom maps
4. API from which proposal engine could access atom map