Add normalized proposal matrix
This commit adds a normalized proposal matrix that can be used by a proposal
engine. Although the actual proposal can be done by the proposal engine, it is
useful to have this functionality in the atom mapper class, since we can check
whether the map scheme is actually leading to a valid set of proposals.