Updating the NCMCEngine to accommodate the new hybrid engine.
Purpose: The old NCMCEngine was using the old absolute factory and the old hybrid.
It also had code duplication and was not using the new OpenMMtools framework.
As a result, the code was more complex than necessary and no longer functions properly.

Updates: This commit begins to remove excess code. Future commits will:

1. Remove all integrators contained here (they are now maintained in OpenMMTools)
2. Use the NonequilibriumSwitchingMove and OpenMMtools objects for integration.
3. Update tests
4. Update absolute example.